+++ packages/expo-sqlite/src/SQLite.ts
@@ -50,7 +50,7 @@ class SQLiteDatabase {
       throw new Error('Unable to delete an opening database');
     }
 
-    return ExponentSQLite.delete(this._name);
+    return ExponentSQLite.deleteAsync(this._name);
   }
 }
 
@@ -113,10 +113,8 @@ export function openDatabase(
     throw new TypeError(`The database name must not be undefined`);
   }
   const db = _openExpoSQLiteDatabase(name, version, description, size, callback);
-  const extendedMethods = ['exec', 'close', 'deleteAsync'];
-  const dbWithExtendedMethods = extendedMethods.reduce((curr, methodName) => {
-    curr[methodName] = curr._db[methodName].bind(curr._db);
-    return curr;
-  }, db);
-  return dbWithExtendedMethods;
+  db.exec = db._db.exec.bind(db._db);
+  db.closeAsync = db._db.close.bind(db_.db);
+  db.deleteAsync = db._db.deleteAsync.bind(db._db);
+  return db;
 }